+++ regionmask/defined_regions/prudence.py
@@ -55,4 +55,5 @@ prudence = Regions(
     abbrevs=short_names,
     name="PRUDENCE",
     source=source,
+    overlap=True,
 )